feat(json): changes json logging to log objects directly to output
Instead of logging objects to the message property, objects passed to Ogma in with the JSON format
as true will be logged directly to the object. This means that if `ogma.log({key1: 'value1, nest:
{key2: 'value2'}})` is logged, the output can have the keys time, pid, application, level, context,
key1, and nest.

BREAKING CHANGE: Objects logged via JSON format are no longer found under the message property